requirements.txt
fastapi
pydantic-settings
loguru
pandas
openpyxl
pillow
openai
python-dotenv
google-generativeai
httpx
